requirements.txt
flask
libretranslate